Refactor handlers to use service layer for URL resolution, add error handling, and update package structure.
Fix issues 1-5
