+++ cmd/shortener/shortener.go
@@ -3,5 +3,5 @@ package main
 import "github.com/nikitachukov/url_shortener.git/internal/app"
 
 func main() {
-	router.StartServer()
+	app.StartServer()
 }
