+++ internal/app/app.go
@@ -1,4 +1,4 @@
-package router
+package app
 
 import (
 	"errors"
